+++ src/Set.php
@@ -462,4 +462,24 @@ final class Set implements \Countable
             },
         );
     }
+
+    /**
+     * Make sure every value conforms to the assertion, you must throw an
+     * exception when a value does not conform.
+     *
+     * For deferred and lazy sets the assertion is called on the go, meaning
+     * subsequent operations may start before reaching a value that doesn't
+     * conform. To be used carefully.
+     *
+     * @template R
+     *
+     * @param R $carry
+     * @param callable(R, T): R $assert
+     *
+     * @return self<T>
+     */
+    public function safeguard($carry, callable $assert)
+    {
+        return new self($this->implementation->safeguard($carry, $assert));
+    }
 }