+++ src/Set/Defer.php
@@ -101,10 +101,6 @@ final class Defer implements Implementation
      */
     public function remove($element): self
     {
-        if (!$this->contains($element)) {
-            return $this;
-        }
-
         return new self($this->values->filter(
             static fn($value) => $value !== $element,
         ));